PROBLEM 5-1
Consider the following code and then answer the coding question and the following two questions after.

code_exam5.py (click to download .py file)

Or click to see code_exam5.py as text

Fill in the code for setGrade, getGrade, setPset, and getPset, using the courseInfo class functions. Paste your entire definition of the edx class in the following box.